refactor: userinfo tabs
