+++ apps/lostark-hands-next/app/markets/@component/goods-list/card.tsx
