+++ apps/lostark-hands-next/app/markets/@component/goods-list/price-label.tsx
