+++ apps/lostark-hands-next/app/markets/@component/index.tsx
@@ -7,11 +7,11 @@ import { Button } from '@sono-repo/ui';
 
 import type { IOptions } from '@/service/markets/types';
 
-import useMarketsFilter from '@/app/markets/@hook/use-markets-filter';
 import useClientRendered from '@/hook/use-client-rendered';
 
 import List from './goods-list';
 import ListSkeleton from './goods-list/skeleton';
+import useMarketsFilter from './use-markets-filter';
 
 interface IMarketsProps {
 	data: IOptions;
