+++ apps/lostark-hands-next/app/user-info/[name]/collection/@component/constants.ts
@@ -1,0 +1,8 @@
+import cn from 'classnames';
+
+export const STICKY_NAV_STYLE = cn(
+	'flex flex-nowrap',
+	'z-[90] mx-[-16px] mb-[8px] self-start px-[16px] pb-[8px]',
+	'hide-scrollbar w-screen overflow-x-scroll',
+	'lg:m-0 lg:mr-[30px] lg:w-fit lg:flex-col lg:overflow-x-visible lg:bg-transparent lg:px-0 lg:pb-0'
+);
