+++ apps/lostark-hands-next/app/user-info/[name]/silblings/page.tsx
@@ -3,10 +3,8 @@ import ServerWrapper from '@/app/server-wrapper';
 import { getSilblingsInfoApi } from '@/service/characters';
 import { sliblingListSelector } from '@/service/characters/selector';
 
-import {
-	Silblings,
-	SilblingsSkeleton
-} from '@/app/user-info/[name]/silblings/@component';
+import Silblings from './@component';
+import SilblingsSkeleton from './@component/skeleton';
 
 const Page = ({ params: { name } }: { params: { name: string } }) => {
 	return (
